requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pytrends==4.9.2
